update code logic

+++ Entities/Subject.cs
@@ -7,8 +7,8 @@ namespace Entities
         [Key]
         public int Id { get; set; }
         public string Name { get; set; }
-        public float MidtermWeight { get; set; }
-        public float FinalWeight { get; set; }
+        public double MidtermWeight { get; set; }
+        public double FinalWeight { get; set; }
         public short Semester { get; set; }
     }
 }
